@@ -1,11 +1,15 @@
-module github.com/shomali11/slacker
+module github.com/shomali11/slacker/v2
 
-go 1.14
+go 1.18
 
 require (
 	github.com/robfig/cron v1.2.0
 	github.com/shomali11/commander v0.0.0-20220716022157-b5248c76541a
 	github.com/shomali11/proper v0.0.0-20180607004733-233a9a872c30
 	github.com/slack-go/slack v0.12.1
+)
+
+require (
+	github.com/gorilla/websocket v1.4.2 // indirect
 	github.com/stretchr/testify v1.3.0 // indirect
 )
