@@ -4,11 +4,11 @@ import "log"
 
 var logDebugMode = false
 
-func infof(format string, v ...interface{}) {
+func infof(format string, v ...any) {
 	log.Printf(format, v...)
 }
 
-func debugf(format string, v ...interface{}) {
+func debugf(format string, v ...any) {
 	if logDebugMode {
 		log.Printf(format, v...)
 	}
