@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/shomali11/slacker"
+	"github.com/shomali11/slacker/v2"
 )
 
 // Defining a command using slacker
@@ -14,8 +14,8 @@ func main() {
 	bot := slacker.NewClient(os.Getenv("SLACK_BOT_TOKEN"), os.Getenv("SLACK_APP_TOKEN"))
 
 	definition := &slacker.CommandDefinition{
-		Handler: func(botCtx slacker.CommandContext) {
-			botCtx.Response().Reply("pong")
+		Handler: func(ctx slacker.CommandContext) {
+			ctx.Response().Reply("pong")
 		},
 	}
 
