@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/shomali11/slacker"
+	"github.com/shomali11/slacker/v2"
 )
 
 // Defining a command with two parameters. Parsing one as a string and the other as an integer.
@@ -17,11 +17,11 @@ func main() {
 	definition := &slacker.CommandDefinition{
 		Description: "Repeat a word a number of times!",
 		Examples:    []string{"repeat hello 10"},
-		Handler: func(botCtx slacker.CommandContext) {
-			word := botCtx.Request().StringParam("word", "Hello!")
-			number := botCtx.Request().IntegerParam("number", 1)
+		Handler: func(ctx slacker.CommandContext) {
+			word := ctx.Request().StringParam("word", "Hello!")
+			number := ctx.Request().IntegerParam("number", 1)
 			for i := 0; i < number; i++ {
-				botCtx.Response().Reply(word)
+				ctx.Response().Reply(word)
 			}
 		},
 	}
