@@ -5,15 +5,15 @@ import (
 	"log"
 	"os"
 
-	"github.com/shomali11/slacker"
+	"github.com/shomali11/slacker/v2"
 	"github.com/slack-go/slack"
 )
 
 func main() {
 	bot := slacker.NewClient(os.Getenv("SLACK_BOT_TOKEN"), os.Getenv("SLACK_APP_TOKEN"))
 
-	bot.UnhandledInteractiveCallback(func(botCtx slacker.InteractiveContext) {
-		callback := botCtx.Callback()
+	bot.UnhandledInteractiveCallback(func(ctx slacker.InteractiveContext) {
+		callback := ctx.Callback()
 		if callback.Type != slack.InteractionTypeBlockActions {
 			return
 		}
@@ -37,20 +37,20 @@ func main() {
 			text = "I don't understand your mood..."
 		}
 
-		_, _, _ = botCtx.APIClient().PostMessage(callback.Channel.ID, slack.MsgOptionText(text, false),
+		_, _, _ = ctx.APIClient().PostMessage(callback.Channel.ID, slack.MsgOptionText(text, false),
 			slack.MsgOptionReplaceOriginal(callback.ResponseURL))
 
-		botCtx.SocketModeClient().Ack(*botCtx.Event().Request)
+		ctx.SocketModeClient().Ack(*ctx.Event().Request)
 	})
 
 	definition := &slacker.CommandDefinition{
-		Handler: func(botCtx slacker.CommandContext) {
+		Handler: func(ctx slacker.CommandContext) {
 			happyBtn := slack.NewButtonBlockElement("happy", "true", slack.NewTextBlockObject("plain_text", "Happy 🙂", true, false))
 			happyBtn.Style = "primary"
 			sadBtn := slack.NewButtonBlockElement("sad", "false", slack.NewTextBlockObject("plain_text", "Sad ☹️", true, false))
 			sadBtn.Style = "danger"
 
-			botCtx.Response().Reply("", slacker.WithBlocks([]slack.Block{
+			ctx.Response().Reply("", slacker.WithBlocks([]slack.Block{
 				slack.NewSectionBlock(slack.NewTextBlockObject(slack.PlainTextType, "What is your mood today?", true, false), nil, nil),
 				slack.NewActionBlock("mood-block", happyBtn, sadBtn),
 			}))
